converted multiple file plugin configs into modules
