--- lua/configs/github-notifications/config.lua
@@ -1,6 +1,0 @@
-local secrets = require 'secrets'
-
-require('github-notifications').setup {
-  username = secrets.username,
-  token = secrets.token,
-}
