--- lua/configs/lsp/settings/bashls.lua
@@ -1,1 +1,0 @@
-return {}
