--- lua/configs/lsp/settings/eslint.lua
@@ -1,1 +1,0 @@
-return {}
