--- lua/configs/lsp/settings/solargraph.lua
@@ -1,1 +1,0 @@
-return {}
