--- lua/configs/lsp/settings/sumneko_lua.lua
@@ -1,15 +1,0 @@
-return {
-  settings = {
-    Lua = {
-      diagnostics = {
-        globals = { 'vim' },
-      },
-      workspace = {
-        library = {
-          [vim.fn.expand '$VIMRUNTIME/lua'] = true,
-          [vim.fn.stdpath 'config' .. '/lua'] = true,
-        },
-      },
-    },
-  },
-}
