--- lua/configs/lsp/settings/tsserver.lua
@@ -1,1 +1,0 @@
-return {}
