--- lua/configs/lsp/settings/vuels.lua
@@ -1,1 +1,0 @@
-return {}
