--- lua/configs/lsp/settings/yamlls.lua
@@ -1,1 +1,0 @@
-return {}
