--- lua/configs/nvim-toggleterm/setup.lua
@@ -1,9 +1,0 @@
-local mappings = {
-  j = {
-    name = 'Terminal',
-    d = { '<cmd>lua _LAZYDOCKER_TOGGLE()<CR>', 'Lazy Docker' },
-    l = { '<cmd>lua _LAZYGIT_TOGGLE()<CR>', 'Lazy Git' },
-  },
-}
-
-require('which-key').register(mappings, { prefix = '<leader>' })
