+++ lua/configs/treesitter-textobjects.lua
@@ -1,0 +1,75 @@
+local M = {}
+
+function M.setup()
+  local mappings = {
+    d = {
+      name = 'Peek',
+      f = { 'Peek Function' },
+      F = { 'Peek Class' },
+    },
+  }
+
+  require('which-key').register(mappings, { prefix = '<leader>' })
+end
+
+function M.config()
+  require('nvim-treesitter.configs').setup {
+    textobjects = {
+      lsp_interop = {
+        enable = true,
+        border = 'rounded',
+        peek_definition_code = {
+          ['<leader>df'] = '@function.outer',
+          ['<leader>dF'] = '@class.outer',
+        },
+      },
+      move = {
+        enable = true,
+        set_jumps = true, -- whether to set jumps in the jumplist
+        goto_next_start = {
+          [']f'] = '@function.outer',
+          [']]'] = '@class.outer',
+        },
+        goto_next_end = {
+          [']F'] = '@function.outer',
+          [']['] = '@class.outer',
+        },
+        goto_previous_start = {
+          ['[f'] = '@function.outer',
+          ['[['] = '@class.outer',
+        },
+        goto_previous_end = {
+          ['[F'] = '@function.outer',
+          ['[]'] = '@class.outer',
+        },
+      },
+      select = {
+        enable = true,
+        -- Automatically jump forward to textobj, similar to targets.vim
+        lookahead = true,
+        keymaps = {
+          ['af'] = '@function.outer',
+          ['if'] = '@function.inner',
+          ['ac'] = '@class.outer',
+          ['ic'] = '@class.inner',
+          ['al'] = '@loop.outer',
+          ['il'] = '@loop.inner',
+          ['aa'] = '@parameter.outer',
+          ['ia'] = '@parameter.inner',
+          ['uc'] = '@comment.outer',
+        },
+      },
+      swap = {
+        enable = false,
+        swap_next = {
+          ['<leader>a'] = '@parameter.inner',
+        },
+        swap_previous = {
+          ['<leader>A'] = '@parameter.inner',
+        },
+      },
+    },
+  }
+end
+
+return M
