--- lua/configs/treesitter-textobjects/setup.lua
@@ -1,9 +1,0 @@
-local mappings = {
-  d = {
-    name = 'Peek',
-    f = { 'Peek Function' },
-    F = { 'Peek Class' },
-  },
-}
-
-require('which-key').register(mappings, { prefix = '<leader>' })
